% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action-helpers.R
\name{action-helpers}
\alias{action-helpers}
\alias{call_addin}
\alias{replace_selection}
\title{Action Helpers}
\usage{
call_addin(package, name)

replace_selection(txt)
}
\arguments{
\item{package}{A string. The addin's package}

\item{name}{A string. The addin's name}

\item{txt}{A character object, or will be coerced to character using \code{deparse()}.
Used as a replacement for the selection text.}
}
\description{
These helpers are meant to be used on the right hand side of the trick specification,
which is evaluated when the trick is selected.
}
\details{
\itemize{
\item \code{replace_selection} replaces the selection in the editor with its argument
\item \code{call_addin} calls a standard RStudio addin
}
}
