% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-helpers.R
\name{condition-helpers}
\alias{condition-helpers}
\alias{selection_is_empty}
\alias{selection_is_comment_line}
\alias{selection_is_comment_block}
\alias{selection_is_n_lines}
\alias{selection_is_single_line}
\alias{selection_is_parsable}
\alias{project_is_package}
\title{Condition Helpers}
\usage{
selection_is_empty(.txt = current_selection())

selection_is_comment_line()

selection_is_comment_block()

selection_is_n_lines(n)

selection_is_single_line()

selection_is_parsable(
  multiline_ok = TRUE,
  single_line_ok = TRUE,
  symbol_ok = TRUE,
  empty_ok = FALSE
)

project_is_package(
  uses_git = NA,
  uses_github = NA,
  uses_c = NA,
  uses_cran_comments = NA
)
}
\arguments{
\item{txt}{A character object}

\item{to}{function names whose calls are to be detected}
}
\description{
These helpers are meant to be used on the left hand side of the trick specification,
to filter which tricks the addin should display.
}
