% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_current_selection.R
\name{current_selection}
\alias{current_selection}
\alias{current_call}
\alias{current_expr}
\alias{current_value}
\alias{current_line_numbers}
\alias{current_lines}
\alias{current_indentation}
\alias{current_path}
\alias{document_code}
\alias{current_code_block}
\alias{current_quoted_status}
\title{Current selection}
\usage{
current_selection(
  target = c("default", "lines", "script"),
  window = c("source", "console", "active")
)

current_call(target = c("default", "lines", "script"))

current_expr(target = c("default", "lines", "script"))

current_value(target = c("default", "lines", "script"))

current_line_numbers(target = c("default", "lines", "script"))

current_lines(target = c("default", "lines", "script"))

current_indentation(target = c("default", "lines", "script"))

current_path(full = TRUE)

document_code()

current_code_block()

current_quoted_status()
}
\arguments{
\item{target}{If \code{target} is \code{"lines"} the selection is extended to lines,
if it is \code{"script"} it is extended to the full script}

\item{window}{One of \code{"source"} (for the source window), \code{"console"}
(for the console) or \code{"active"} (for the active document).
See \link{rstudio-editors}.}

\item{full}{boolean. Whether to return full path or base name}
}
\description{
Utilities to get data about the selection
}
