% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-yaml.R
\name{edit_yaml_tricks}
\alias{edit_yaml_tricks}
\title{Edit YAML tricks}
\usage{
edit_yaml_tricks(project_level = FALSE)
}
\arguments{
\item{project_level}{whether to open the project level file ("./.r-tricks.yaml"),
by default opens the user level file ("~/.r-tricks.yaml")}
}
\value{
Returns the input invisibly, called for side effects.
}
\description{
Edit YAML tricks
}
