% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_utils.R
\name{fails}
\alias{fails}
\title{Test if a Call Fails}
\usage{
fails(call)
}
\arguments{
\item{call}{A call}
}
\description{
This is mainly used internally and should be rarely needed but can be useful
for some custom trick conditions. Use this with caution! The tested call is run
so you don't want it to have potential side effects or be slow!
}
\details{
The potential failure will be totally silent (no printed output, no error, no message, no warning).
Other side effects should not be an issues if the function is used properly.
}
