% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_current_file-condition-helpers.R
\name{current_file_is_r_script}
\alias{current_file_is_r_script}
\alias{current_file_is_rmd}
\title{File Focused Condition Helpers}
\usage{
current_file_is_r_script()

current_file_is_rmd()
}
\description{
File Focused Condition Helpers
}
