% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-yaml.R
\name{install_tricks}
\alias{install_tricks}
\title{Install tricks from a package or path}
\usage{
install_tricks(source = "tricks", project_level = FALSE, all = FALSE)
}
\arguments{
\item{source}{The name of a package containing tricks., or a path to a YAML
file defining tricks, By default proposes tricks from the \pkg{tricks} package}

\item{project_level}{whether to install tricks at the project level (in "./.r-tricks.yaml"),
by default they're installed at the user level (in "~/.r-tricks.yaml")}

\item{all}{whether to install all tricks without prompting}
}
\value{
Returns the first argument invisibly, called for side effects.
}
\description{
Install tricks from a package or path
}
