% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tricks.R
\name{load_tricks}
\alias{load_tricks}
\alias{unload_tricks}
\alias{uninstall_tricks}
\alias{loaded_tricks}
\alias{edit_tricks}
\title{Add, Remove or Show Tricks}
\usage{
load_tricks(..., .reset = FALSE)

unload_tricks(...)

uninstall_tricks(..., project_level = FALSE)

loaded_tricks()

edit_tricks(project_level = FALSE)
}
\arguments{
\item{...}{tricks to add (or replace), or names of tricks to remove}

\item{.reset}{whether to remove existing tricks}

\item{project_level}{whether to open the project level file ("./.r-tricks.yaml"),
by default opens the user level file ("~/.r-tricks.yaml")}
}
\description{
Add
}
