% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-yaml.R
\name{load_yaml_tricks}
\alias{load_yaml_tricks}
\title{Load tricks from YAML file}
\usage{
load_yaml_tricks(file = NULL, labels = NULL, reset = FALSE)
}
\arguments{
\item{file}{Path to a YAML file, if \code{NULL} (the default) tricks are loaded
from the user level and project level \code{.r-tricks.yaml} files if they exist}

\item{labels}{labels of tricks to load, by default all tricks are loaded}

\item{reset}{Whether to unload previously loaded tricks}
}
\value{
Returns the first argument invisibly, called for side effects.
}
\description{
Load tricks from YAML file
}
\examples{
yaml_path <- system.file("tricks.yaml", package = "tricks")
load_yaml_tricks(yaml_path)
}
