% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-mock.R
\name{mock_context}
\alias{mock_context}
\title{Mock context to test tricks}
\usage{
mock_context(contents = "", id = 1, path = "", selection = NULL)
}
\arguments{
\item{contents}{A character vector containing lines of code}

\item{id}{A number}

\item{path}{A string}

\item{selection}{A nested list where elements are named lists of \code{range} created
by \code{rstudioapi::document_range()} and \code{position} created by \code{rstudioapi::document_range()}.
By default \code{contents} is fully selected, which should be enough for most tests.}
}
\value{
Returns \code{NULL} invisibly, Called for side effects.
}
\description{
\code{mock_context()} does some magic so calls to \code{current_selection()} and other
functions using context information (i.e. document content, id, path, and selection(s))
will fetch preset data. It is useful to design tests. call \code{unmock}
}
