% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_project-condition-helpers.R
\name{project_is_package}
\alias{project_is_package}
\alias{project_uses_git}
\title{Project Focused Condition Helpers}
\usage{
project_is_package()

project_uses_git()
}
\description{
Project Focused Condition Helpers
}
