% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raven_scales.R
\name{raven_scales}
\alias{raven_scales}
\title{Compute series (A, B, C, D, E) and total score for Raven's Standard Progressive Matrices}
\usage{
raven_scales(data)
}
\arguments{
\item{data}{60-column matrix/data.frame}
}
\value{
matrix of the size \code{nrow(data)} x 6
}
\description{
Compute series (A, B, C, D, E) and total score for Raven's Standard Progressive Matrices
}
\examples{
\dontrun{
D <- cbind(D, raven_scales(D))
}
}
