% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_utils.R
\name{save_by_format}
\alias{save_by_format}
\title{Save object by format}
\usage{
save_by_format(
  object,
  fmt,
  file_path,
  scale,
  width,
  height,
  units,
  dpi,
  engine,
  verbose,
  ...
)
}
\arguments{
\item{object}{The object to save.}

\item{fmt}{Character string specifying the desired file format. E.g. \code{"png"},\code{"pdf"}, \code{"rds"}.}

\item{file_path}{Full file path where the object should be saved.}

\item{scale}{Numeric scaling factor for width and height. Passed to ggsave when used.}

\item{width}{Width of the output in units specified by \code{units}.}

\item{height}{Height of the output in units specified by \code{units}.}

\item{units}{Character string specifying the units for \code{width} and \code{height}.
Options include \code{"in"}, \code{"cm"}, \code{"mm"}, or \code{"pt"}.}

\item{dpi}{Resolution in dots per inch, relevant for raster output.}

\item{engine}{Character string indicating the rendering engine to use. For instance,
\code{"gg"} will attempt to use \code{ggsave} for ggplot objects.}

\item{verbose}{Logical; if \code{TRUE}, enables debugging messages such as the selected engine
and additional processing information.}

\item{...}{Additional arguments passed to the underlying save function.}
}
\value{
No return value; the function is called for its side effects.
}
\description{
Save an object to a specific file format
Saves the object in the specified format using the appropriate saving method.
Uses basic plotting functions for general objects, and may call specialised
functions for recognised object types depending on the chosen engine.
}
