% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_hat.R
\name{sorting_hat}
\alias{sorting_hat}
\title{Gather information about the data variables}
\usage{
sorting_hat(df)
}
\arguments{
\item{df}{The input data frame}
}
\description{
Gather information about the data variables
}
