% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-yaml.R
\name{use_yaml_tricks}
\alias{use_yaml_tricks}
\title{Use YAML tricks}
\usage{
use_yaml_tricks(project_level = FALSE)
}
\arguments{
\item{project_level}{Whether to setup tricks at the project level}
}
\value{
Returns the input invisibly, called for side effects.
}
\description{
Sets up YAML tricks for user or project
}
\details{
\itemize{
\item Edits user level or project level ".RProfile" to add a call to \code{load_yaml_tricks()}
\item Optionally runs installation of tricks from \pkg{tricks} package (default for user level YAML)
\item Adds the project level \code{.r-tricks.yaml} file to \code{.Rbuildignore}
}
}
