% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-yaml.R
\name{yaml_to_trick_list}
\alias{yaml_to_trick_list}
\alias{trick_list_to_yaml}
\title{Read YAML file into a list of trick objects and back}
\usage{
yaml_to_trick_list(file)

trick_list_to_yaml(tricks, file, append = TRUE)
}
\arguments{
\item{file}{Path to a YAML file defining tricks}

\item{tricks}{list of tricks to write to \code{file}}

\item{append}{Whether to append to the \code{file}, if \code{FALSE} the file will be overwrtitten.}
}
\value{
Returns the first argument invisibly, called for side effects.
}
\description{
You should rarely need to manipulate trick objects. A use case is creating a
library of tricks from an existing
}
\examples{
yaml_path <- system.file("tricks.yaml", package = "tricks")
tricks <- yaml_to_trick_list(yaml_path)
labels <- c("Edit user '.Rprofile'","Edit project '.Rprofile'")
trick_subset <- tricks[labels]
new_trick_library_path <- tempfile(fileext = ".yaml")
trick_list_to_yaml(trick_subset, new_trick_library_path)
}
