% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricks.R
\name{add_tricks}
\alias{add_tricks}
\alias{rm_tricks}
\alias{show_tricks}
\title{Add, Remove or Show Tricks}
\usage{
add_tricks(...)

rm_tricks(...)

show_tricks()
}
\arguments{
\item{...}{tricks to add (or replace), or names of tricks to remove}
}
\description{
Add, Remove or Show Tricks
}
