% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_clipboard-condition-helpers.R
\name{clipboard_contains_text}
\alias{clipboard_contains_text}
\alias{clipboard_is_parsable}
\alias{clipboard_text}
\title{Clipboard Focused Condition Helpers}
\usage{
clipboard_contains_text()

clipboard_is_parsable()

clipboard_text()
}
\description{
Clipboard Focused Condition Helpers
}
