% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-helpers.R
\name{condition-helpers}
\alias{condition-helpers}
\alias{is_comment_line}
\alias{is_comment_block}
\alias{is_parsable}
\alias{is_call}
\title{Condition Helpers}
\usage{
is_comment_line(txt)

is_comment_block(txt)

is_parsable(txt)

is_call(txt, to = NULL)
}
\arguments{
\item{txt}{A character object}

\item{to}{function names whose calls are to be detected}
}
\description{
These helpers are meant to be used on the left hand side of the trick specification,
to filter which tricks the addin should display.
}
