% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{current_selection}
\alias{current_selection}
\alias{current_call}
\alias{current_expr}
\alias{current_value}
\alias{current_line_numbers}
\alias{current_lines}
\alias{current_path}
\alias{current_file_code}
\alias{current_code_block}
\title{Current selection}
\usage{
current_selection(target = c("default", "lines", "file"))

current_call(target = c("default", "lines", "file"))

current_expr(target = c("default", "lines", "file"))

current_value(target = c("default", "lines", "file"))

current_line_numbers(target = c("default", "lines", "file"))

current_lines(target = c("default", "lines", "file"))

current_path(full = TRUE)

current_file_code()

current_code_block()
}
\arguments{
\item{full}{boolean. Whether to return full path or base name}
}
\description{
Utilities to get data about the selection
}
