% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{current_selection}
\alias{current_selection}
\alias{current_call}
\alias{current_expr}
\alias{current_value}
\alias{current_line_numbers}
\alias{current_lines}
\alias{current_path}
\alias{current_file_code}
\alias{current_code_block}
\title{Current selection}
\usage{
current_selection()

current_call()

current_expr()

current_value()

current_line_numbers()

current_lines()

current_path(full = TRUE)

current_file_code()

current_code_block()
}
\arguments{
\item{full}{boolean. Whether to return full path or base name}
}
\description{
Utilities to get data about the selection
}
