% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_current_file-condition-helpers.R
\name{document_is_r_script}
\alias{document_is_r_script}
\alias{document_is_rmd}
\title{File Focused Condition Helpers}
\usage{
document_is_r_script()

document_is_rmd()
}
\description{
File Focused Condition Helpers
}
