% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trick.R
\name{new_trick}
\alias{new_trick}
\title{Create a new trick}
\usage{
new_trick(label, condition, action, description = label, ...)
}
\arguments{
\item{label}{A string}

\item{condition}{A call or a one sided formula, using condition helpers and context informers, whose
right hand side expressions returns a boolean when evaluated by {tricks}}

\item{action}{A call or a one sided formula, using action helpers and context informers, whose
right hand side triggers the desired actions when evaluated by {tricks}}

\item{description}{The description of the trick, for documentation purposes}

\item{...}{other properties to print with the object and in YAML files}
}
\value{
A list of class "trick_object"
}
\description{
Create a trick object
}
\examples{
new_trick(
"Edit user '.Rprofile'",
~ selection_is_empty(),
~ usethis::edit_r_profile()
)
}
